.github/actions/atlas-cfn-deploy/requirements.txt
loguru
pyyaml
plac
# for s3wipe
boto
